improvements

+++ test/__init__.py
@@ -191,6 +191,7 @@ class ClientContext(object):
         self.sessions_enabled = False
         self.client = None
         self.conn_lock = threading.Lock()
+        self.data_lake = False
 
         if COMPRESSORS:
             self.default_client_options["compressors"] = COMPRESSORS
@@ -231,6 +232,15 @@ class ClientContext(object):
 
     def _init_client(self):
         self.client = self._connect(host, port)
+
+        build_info = self.client.admin.command('buildInfo')
+        if 'dataLake' in build_info:
+            self.data_lake = True
+            self.client = self._connect(
+                host, port, username=db_user, password=db_pwd)
+            self.connected = True
+            return
+
         if HAVE_SSL and not self.client:
             # Is MongoDB configured for SSL?
             self.client = self._connect(host, port, **TLS_OPTIONS)
@@ -845,7 +855,7 @@ def teardown():
     if garbage:
         assert False, '\n'.join(garbage)
     c = client_context.client
-    if c:
+    if c and not client_context.data_lake:
         c.drop_database("pymongo-pooling-tests")
         c.drop_database("pymongo_test")
         c.drop_database("pymongo_test1")
